+++ tests/node-traversal.rs
@@ -158,6 +158,31 @@ fn test_descendants_bound() {
     assert_eq!(no_descendants_node.descendants_it().count(), 0);
 }
 
+
+#[cfg_attr(not(target_arch = "wasm32"), test)]
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+fn test_descendants_after_mod() {
+    // previously `DescendantNodes` could traverse beyond the initial node when iterating over descendants.
+    let doc: Document = ANCESTORS_CONTENTS.into();
+
+    let parent = doc.select_single("#parent");
+    let parent_node = parent.nodes().first().unwrap();
+
+    let grand_parent = doc.select_single("#grand-parent");
+    let grand_parent_node = grand_parent.nodes().first().unwrap();
+
+    grand_parent_node.replace_with(parent_node);
+    parent_node.append_child(grand_parent_node);
+
+    let descendants_id_names: Vec<String> = parent_node
+        .descendants_it()
+        .filter(|n| n.is_element())
+        .map(|n| n.attr_or("id", "").to_string())
+        .collect();
+    let expected_id_names = vec!["first-child", "second-child", "grand-parent"];
+    assert_eq!(descendants_id_names, expected_id_names);
+}
+
 #[cfg_attr(not(target_arch = "wasm32"), test)]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
 fn test_last_child() {